Fix filtering inconsistencies heatmap and search climbs
Abstract out all the filtering logic to a new abstraction, and
re-use that abstraction in both the heatmap and search maps queries.
This ensures that the heatmap is actually correct compared to the search
climbs result.
