+++ app/api/v1/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/search/route.ts
@@ -1,4 +1,4 @@
-import { searchClimbs } from '@/app/lib/db/queries/search-climbs';
+import { searchClimbs } from '@/app/lib/db/queries/climbs/search-climbs';
 import { BoardRouteParameters, ErrorResponse, SearchClimbsResult, SearchRequestPagination } from '@/app/lib/types';
 import { parseBoardRouteParams, urlParamsToSearchParams } from '@/app/lib/url-utils';
 import { NextResponse } from 'next/server';
