+++ app/lib/db/queries/climbs/create-climb-filters.ts
@@ -1,0 +1,119 @@
+import { and, eq, between, gte, sql, like, notLike, SQL } from 'drizzle-orm';
+import { alias, PgTableWithColumns } from 'drizzle-orm/pg-core';
+import { ParsedBoardRouteParameters, SearchRequestPagination } from '@/app/lib/types';
+import { TableSet } from '@/lib/db/queries/util/table-select';
+
+/**
+ * Creates a shared filtering object that can be used by both search climbs and heatmap queries
+ * @param tables The board-specific tables from getBoardTables
+ * @param params The route parameters
+ * @param searchParams The search parameters
+ * @param productSizeAlias Optional alias for the product sizes table. If provided, size conditions will use this alias.
+ */
+export const createClimbFilters = (
+  tables: TableSet,
+  params: ParsedBoardRouteParameters,
+  searchParams: SearchRequestPagination,
+  productSizeAlias?: PgTableWithColumns<any>,
+) => {
+  // Process hold filters
+  const holdsToFilter = Object.entries(searchParams.holdsFilter || {}).map(([key, state]) => [
+    key.replace('hold_', ''),
+    state,
+  ]);
+
+  const anyHolds = holdsToFilter.filter(([, value]) => value === 'ANY').map(([key]) => Number(key));
+  const notHolds = holdsToFilter.filter(([, value]) => value === 'NOT').map(([key]) => Number(key));
+
+  // Base conditions for filtering climbs that don't reference the product sizes table
+  const baseConditions: SQL[] = [
+    eq(tables.climbs.layoutId, params.layout_id),
+    eq(tables.climbs.isListed, true),
+    eq(tables.climbs.isDraft, false),
+    eq(tables.climbs.framesCount, 1),
+  ];
+
+  // Get the product sizes table or its alias
+  const sizeTable = productSizeAlias || tables.productSizes;
+
+  // Size-specific conditions
+  const sizeConditions: SQL[] = [
+    eq(sizeTable.id, params.size_id),
+    sql`${tables.climbs.edgeLeft} > ${sizeTable.edgeLeft}`,
+    sql`${tables.climbs.edgeRight} < ${sizeTable.edgeRight}`,
+    sql`${tables.climbs.edgeBottom} > ${sizeTable.edgeBottom}`,
+    sql`${tables.climbs.edgeTop} < ${sizeTable.edgeTop}`,
+  ];
+
+  // Conditions for climb stats
+  const climbStatsConditions: SQL[] = [];
+
+  if (searchParams.minAscents) {
+    climbStatsConditions.push(gte(tables.climbStats.ascensionistCount, searchParams.minAscents));
+  }
+
+  if (searchParams.minGrade && searchParams.maxGrade) {
+    climbStatsConditions.push(
+      sql`ROUND(${tables.climbStats.displayDifficulty}::numeric, 0) BETWEEN ${searchParams.minGrade} AND ${searchParams.maxGrade}`,
+    );
+  } else if (searchParams.minGrade) {
+    climbStatsConditions.push(
+      sql`ROUND(${tables.climbStats.displayDifficulty}::numeric, 0) >= ${searchParams.minGrade}`,
+    );
+  } else if (searchParams.maxGrade) {
+    climbStatsConditions.push(
+      sql`ROUND(${tables.climbStats.displayDifficulty}::numeric, 0) <= ${searchParams.maxGrade}`,
+    );
+  }
+
+  if (searchParams.minRating) {
+    climbStatsConditions.push(sql`${tables.climbStats.qualityAverage} >= ${searchParams.minRating}`);
+  }
+
+  if (searchParams.gradeAccuracy) {
+    climbStatsConditions.push(
+      sql`ABS(ROUND(${tables.climbStats.displayDifficulty}::numeric, 0) - ${tables.climbStats.difficultyAverage}::numeric) <= ${searchParams.gradeAccuracy}`,
+    );
+  }
+
+  // Name search condition (only used in searchClimbs)
+  const nameCondition: SQL[] = searchParams.name ? [sql`${tables.climbs.name} ILIKE ${`%${searchParams.name}%`}`] : [];
+
+  // Hold filter conditions
+  const holdConditions: SQL[] = [
+    ...anyHolds.map((holdId) => like(tables.climbs.frames, `%${holdId}r%`)),
+    ...notHolds.map((holdId) => notLike(tables.climbs.frames, `%${holdId}r%`)),
+  ];
+
+  return {
+    // Helper function to get all climb filtering conditions
+    getClimbWhereConditions: () => [...baseConditions, ...nameCondition, ...holdConditions],
+
+    // Size-specific conditions
+    getSizeConditions: () => sizeConditions,
+
+    // Helper function to get all climb stats conditions
+    getClimbStatsConditions: () => climbStatsConditions,
+
+    // For use in the subquery with left join
+    getClimbStatsJoinConditions: () => [
+      eq(tables.climbStats.climbUuid, tables.climbs.uuid),
+      eq(tables.climbStats.angle, params.angle),
+    ],
+
+    // For use in getHoldHeatmapData
+    getHoldHeatmapClimbStatsConditions: (climbHoldsTable: typeof tables.climbHolds) => [
+      eq(tables.climbStats.climbUuid, climbHoldsTable.climbUuid),
+      eq(tables.climbStats.angle, params.angle),
+    ],
+
+    // Raw parts, in case you need direct access to these
+    baseConditions,
+    climbStatsConditions,
+    nameCondition,
+    holdConditions,
+    sizeConditions,
+    anyHolds,
+    notHolds,
+  };
+};
